next_logger/requirements.txt
PySide6>=6.7
pyserial>=3.5
